[Bug Fix]Use BigDecimal Instead of float
When i start adding really large numbers or divide really small numbers,
the prescion of the answer becomes incorrect.
This is because float is capped to a max and min value.
To fix this i change from type float to BigDecimal instead, as
BigDecimal is capable of dealing with really large numbers and small
numbers.

+++ calculator/src/main/java/entry/FactoryEntry.java
@@ -1,5 +1,6 @@
 package entry;
 
+import java.math.BigDecimal;
 import enums.Symbol;
 
 /**
@@ -19,7 +20,7 @@ public class FactoryEntry {
    *
    * @param val The number the {@code Entry} class will hold
    */
-  public NumEntry createEntry(float val) {
+  public NumEntry createEntry(BigDecimal val) {
     return new NumEntry(val);
   }
