+++ calculator/src/main/java/model/CalcModel.java
@@ -1,6 +1,7 @@
 package model;
 
 import enums.OpType;
+import java.math.BigDecimal;
 
 /**
  * A state class which switches between postfix and infix.
@@ -41,7 +42,7 @@ public class CalcModel {
   /**
    * Evaluates user expression either in infix or postfix.
    */
-  public float evaluate(String expression) {
+  public BigDecimal evaluate(String expression) {
     return currentState.evaluate(expression);
   }
 }
