+++ calculator/src/main/java/model/Calculator.java
@@ -1,5 +1,7 @@
 package model;
 
+import java.math.BigDecimal;
+
 /**
  * The {@code Calculator} interface gives user control over what type of 
  * calculation to do with {@code String} expressions. 
@@ -12,7 +14,7 @@ public interface Calculator {
    * Returns a solution for an given expression.
    *
    * @param expression The expression to be calculated 
-   * @return a float which represents the solution to the {@code String} expression
+   * @return a decimal which represents the solution to the {@code String} expression
    */
-  public float evaluate(String expression);
+  public BigDecimal evaluate(String expression);
 }
